Fix unicode port conversion
All key/value pairs from the config are converted to unicode but
httplib does not accept unicode for the port; only number or
string. This converts the host string to unicode by combining
the host and port given (u"example.com:1443") and allowing httplib
to parse the port out of the unicode.